Update unit_transport_preserve_queue.lua
If you have regular unit selected and target the transport (instead of other way around), the last command (in this case order to load the unit) persists and the first thing the unit does after being unloaded is to ask to be immediately loaded again, resulting in endless cycle.

You were talking about automatically removing all move commands at the front of the command que. Looks like having the script remove the load command solves this.